# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tutor Marketplace CRM & Lead Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parents:
    get:
      operationId: listParents
      tags: [parents]
      summary: List all parents
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of parents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parent"
    post:
      operationId: createParent
      tags: [parents]
      summary: Register a parent (lead)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParentInput"
      responses:
        "201":
          description: Parent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
  /parents/{id}:
    get:
      operationId: getParent
      tags: [parents]
      summary: Get parent details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentDetail"
        "404":
          description: Not found
    patch:
      operationId: updateParent
      tags: [parents]
      summary: Update parent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParentUpdate"
      responses:
        "200":
          description: Updated parent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"

  # Leads
...
  /requirements:
    get:
      operationId: listRequirements
      tags: [requirements]
      summary: List tuition requirements (public board)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
        - name: grade
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Requirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Requirement"
    post:
      operationId: createRequirement
      tags: [requirements]
      summary: Create a tuition requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequirementInput"
      responses:
        "201":
          description: Requirement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"
  /requirements/{id}:
    get:
      operationId: getRequirement
      tags: [requirements]
      summary: Get requirement details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequirementDetail"
        "404":
          description: Not found
    patch:
      operationId: updateRequirement
      tags: [requirements]
      summary: Update requirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequirementUpdate"
      responses:
        "200":
          description: Updated requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requirement"

  # Applications
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # Staff
...
components:
  schemas:
...
    Parent:
      type: object
      required: [id, name, phone, email, area, leadStatus, leadSource, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        area:
          type: string
        city:
          type: string
        leadStatus:
          type: string
          enum: [new, contacted, qualified, demo_scheduled, demo_done, converted, lost]
        leadSource:
          type: string
          enum: [website, referral, social_media, walk_in, phone, other]
        assignedTo:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ParentDetail:
      allOf:
        - $ref: "#/components/schemas/Parent"
        - type: object
          properties:
            requirements:
              type: array
              items:
                $ref: "#/components/schemas/Requirement"
            demos:
              type: array
              items:
                $ref: "#/components/schemas/Demo"
            payments:
              type: array
              items:
                $ref: "#/components/schemas/Payment"
    ParentInput:
      type: object
      required: [name, phone, area, leadSource]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        area:
          type: string
        city:
          type: string
        leadSource:
          type: string
          enum: [website, referral, social_media, walk_in, phone, other]
        notes:
          type: string
    ParentUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        area:
          type: string
        city:
          type: string
        leadStatus:
          type: string
          enum: [new, contacted, qualified, demo_scheduled, demo_done, converted, lost]
        leadSource:
          type: string
          enum: [website, referral, social_media, walk_in, phone, other]
        notes:
          type: string
        assignedTo:
          type: ["integer", "null"]
...
    Requirement:
      type: object
      required: [id, parentId, subject, grade, area, status, createdAt]
      properties:
        id:
          type: integer
        parentId:
          type: integer
        parentName:
          type: string
        subject:
          type: string
        grade:
          type: string
        area:
          type: string
        daysPerWeek:
          type: ["integer", "null"]
        duration:
          type: ["integer", "null"]
        budget:
          type: ["number", "null"]
        teachingMode:
          type: string
          enum: [home, online, both]
        gender:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, fulfilled, closed]
        slug:
          type: string
        createdAt:
          type: string
    RequirementDetail:
      allOf:
        - $ref: "#/components/schemas/Requirement"
        - type: object
          properties:
            applications:
              type: array
              items:
                $ref: "#/components/schemas/Application"
    RequirementInput:
      type: object
      required: [parentId, subject, grade, area, teachingMode]
      properties:
        parentId:
          type: integer
        subject:
          type: string
        grade:
          type: string
        area:
          type: string
        daysPerWeek:
          type: integer
        duration:
          type: integer
        budget:
          type: number
        teachingMode:
          type: string
          enum: [home, online, both]
        gender:
          type: string
        description:
          type: string
    RequirementUpdate:
      type: object
      properties:
        subject:
          type: string
        grade:
          type: string
        area:
          type: string
        daysPerWeek:
          type: integer
        duration:
          type: integer
        budget:
          type: number
        teachingMode:
          type: string
          enum: [home, online, both]
        gender:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, fulfilled, closed]
    Application:
      type: object
      required: [id, requirementId, tutorId, status, createdAt]
      properties:
        id:
          type: integer
        requirementId:
          type: integer
        tutorId:
          type: integer
        tutorName:
          type: string
        tutorPhone:
          type: string
        status:
          type: string
          enum: [pending, shortlisted, rejected, selected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Demo:
      type: object
      required: [id, requirementId, tutorId, scheduledAt, status, createdAt]
      properties:
        id:
          type: integer
        requirementId:
          type: integer
        tutorId:
          type: integer
        tutorName:
          type: string
        parentId:
          type: integer
        parentName:
          type: string
        parentPhone:
          type: string
        subject:
          type: string
        area:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
        feedback:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, requirementId, amount, type, status, createdAt]
      properties:
        id:
          type: integer
        requirementId:
          type: integer
        tutorId:
          type: ["integer", "null"]
        parentId:
          type: integer
        parentName:
          type: string
        tutorName:
          type: ["string", "null"]
        subject:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [registration, monthly, demo_fee, other]
        status:
          type: string
          enum: [pending, paid, overdue]
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [requirementId, parentId, amount, type]
      properties:
        requirementId:
          type: integer
        parentId:
          type: integer
        tutorId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [registration, monthly, demo_fee, other]
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, overdue]
        paidAt:
          type: string
        amount:
          type: number
        notes:
          type: string