# lib/api-zod/src/generated/types/requirement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tutor Marketplace CRM & Lead Management API
 * OpenAPI spec version: 0.1.0
 */
import type { RequirementStatus } from './requirementStatus';
import type { RequirementTeachingMode } from './requirementTeachingMode';

export interface Requirement {
  id: number;
  parentId: number;
  parentName?: string;
  subject: string;
  grade: string;
  area: string;
  /** @nullable */
  daysPerWeek?: number | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  budget?: number | null;
  teachingMode?: RequirementTeachingMode;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  description?: string | null;
  status: RequirementStatus;
  slug?: string;
  createdAt: string;
}